+++ Sources/TableViewModelController.swift
@@ -106,6 +106,8 @@ open class TableViewModelController: UITableViewController, ViewModelCellDelegat
             tableView.register(TableCellTextInput.self, forCellReuseIdentifier: identifier)
         case .slider:
             tableView.register(TableCellSlider.self, forCellReuseIdentifier: identifier)
+        case .sliderLabels:
+            tableView.register(TableCellSliderLabels.self, forCellReuseIdentifier: identifier)
         case .customClass(let cellClass):
             tableView.register(cellClass, forCellReuseIdentifier: identifier)
         case .customNib(let cellClass):